requirements.txt
astroid==2.4.2
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-PyMongo==2.3.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
passlib==1.7.4
pylint==2.6.0
pymongo==3.11.1
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
